221124_3
14.05

+++ src/main/java/ezen/store/controller/Mb_Controller.java
@@ -2,6 +2,8 @@ package ezen.store.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -63,6 +65,7 @@ public class Mb_Controller {
 	// 로그인 기능
 	@PostMapping("/Mbloginpro")
 	public String Mbloginpro(@Validated@ModelAttribute("tempMbBean") Mb_Bean tempMbBean, 
+							 HttpSession session,
 							 BindingResult result) {
 
 		if (result.hasErrors()) {
@@ -78,15 +81,32 @@ public class Mb_Controller {
 
 		if (loginCheck == true) {
 			
+			if(session.getAttribute("loginBean") != null) {
+				
+				session.removeAttribute("loginBean");
+				
+				session.setAttribute("loginBean", tempMbBean);
+				
+				
+				
+				return "member/Mb_login_success";
+				
+			}
+			
+			session.setAttribute("loginBean", tempMbBean);
+			
 			return "member/Mb_login_success";
 			
-		} else {
+			} else {
+				
+				tempMbBean.setMblogin(false);
+				
+				return "member/Mb_login";
+			}
+		
+		
 			
-			tempMbBean.setMblogin(loginCheck);
-			
-			return "member/Mb_login_fail";
-			
-		}
+		
 	}
 	
 	// 회원가입 컨트롤
@@ -120,42 +140,53 @@ public class Mb_Controller {
 	}
 	
 
-	// 회원정보 수정 기능
-	@PostMapping("/Mbupdatepro")
-	public String Mbupdatepro(@Validated @ModelAttribute("updateMbBean") Mb_Bean updateMbBean, BindingResult result) {
+		// 회원정보 수정 기능
+		@PostMapping("/Mbupdatepro")
+		public String Mbupdatepro(@Validated@ModelAttribute("updateMbBean") Mb_Bean updateMbBean, BindingResult result) {
 
-		if (result.hasErrors()) {
-			return "member/Mb_update";
+			if (result.hasErrors()) {
+				return "member/Mb_update";
+			}
+
+			// 회원정보 수정 성공페이지
+			mbService.modifyUserInfo(updateMbBean);
+
+			return "member/Mb_update_success";
 		}
 
-		// 회원정보 수정 성공페이지
-		mbService.modifyUserInfo(updateMbBean);
-
-		return "member/Mb_update_success";
-	}
-
 	
 	
-		// 회원정보 삭제 정보 컨트롤 (원본)
+		// 회원정보 삭제 페이지
 		@GetMapping("/Mbdelete")
-		public String Mbdelete(@ModelAttribute("deleteMbBean") Mb_Bean deleteMbBean) {
+		public String Mbdelete(@RequestParam("mb_id") String mb_id, Model model) {
 
-			mbService.getMbInfo(deleteMbBean);
+			Mb_Bean deleteMbBean = mbService.getMbInfo(mb_id);
+			model.addAttribute("deleteMbBean", deleteMbBean);
 
 			return "member/Mb_delete";
 		}
 		
-		//삭제 컨트롤 프로
+		
+		// 회원정보 삭제(처리) 기능 
 		@PostMapping("/Mbdeletepro")
-		public String Mbdeletepro(@Validated@ModelAttribute("deleteMbBean") Mb_Bean deleteMbBean, BindingResult result) {
+		public String Mbdeletepro(@ModelAttribute("deleteMbBean") Mb_Bean deleteMbBean, BindingResult result) {
 			
 			if(result.hasErrors()) {
 				return "member/Mb_delete";
 			}
 			
-			mbService.deleteUserInfo(deleteMbBean);
+			if(deleteMbBean.getMb_pw().equals(deleteMbBean.getMb_pw2())) {
+				
+				mbService.deleteUserInfo(deleteMbBean);
 			
-			return "member/Mb_delete_success";
+				return "member/Mb_delete_success";
+			
+			} else {
+				
+				return "member/Mb_delete_fail";
+				
+			}
+			
 		}
